requirements.txt
streamlit==1.28.1
python-control==0.9.4
numpy==1.24.3
scipy==1.11.0
matplotlib==3.7.2
pandas==2.0.3
plotly==5.14.0
